split tft logic into app and game
