+++ main.py
@@ -3,13 +3,9 @@ The main class for Alune, responsible for the main loop.
 """
 
 import asyncio
-from dataclasses import dataclass
-from enum import auto
-from enum import StrEnum
 import importlib.metadata
 import json
 import os
-from random import Random
 import sys
 from urllib.error import HTTPError
 from urllib.error import URLError
@@ -17,264 +13,13 @@ import urllib.request
 
 from adb_shell.exceptions import TcpTimeoutException
 import google_play_scraper
-import keyboard
 from loguru import logger
-from numpy import ndarray
 
 from alune import helpers
-from alune import screen
 from alune.adb import ADB
 from alune.config import AluneConfig
 from alune.helpers import raise_and_exit
-from alune.images import Button
-from alune.images import Image
-from alune.screen import BoundingBox
-from alune.screen import ImageSearchResult
-
-PAUSE_LOGIC = False
-PLAY_NEXT_GAME = True
-
-
-class GameState(StrEnum):
-    """
-    State the game or app is in.
-    """
-
-    LOADING = auto()
-    MAIN_MENU = auto()
-    CHOOSE_MODE = auto()
-    LOBBY = auto()
-    QUEUE_MISSED = auto()
-    IN_GAME = auto()
-    POST_GAME_DAWN_OF_HEROES = auto()
-    POST_GAME = auto()
-    CHOICE_CONFIRM = auto()
-
-
-@dataclass
-class GameStateImageResult:
-    """
-    Combines a game state with an image search result (both optional)
-    """
-
-    game_state: GameState
-    image_result: ImageSearchResult | None = None
-
-
-async def wait_for_accept_button(adb_instance: ADB):
-    """
-    Utility method to wait for the queue accept button.
-
-    Args:
-        adb_instance: An instance of the ADB connection to click in.
-    """
-    screenshot = await adb_instance.get_screen()
-    search_result = screen.get_button_on_screen(screenshot, Button.accept)
-    while not search_result:
-        await asyncio.sleep(2)
-        screenshot = await adb_instance.get_screen()
-        search_result = screen.get_button_on_screen(screenshot, Button.accept)
-
-
-async def queue(adb_instance: ADB, config: AluneConfig):
-    """
-    Utility method to queue a match.
-
-    Args:
-        adb_instance: An instance of the ADB connection to click in.
-        config: An instance of the alune config to use.
-    """
-    try:
-        await asyncio.wait_for(wait_for_accept_button(adb_instance), timeout=config.get_queue_timeout())
-    except asyncio.TimeoutError:
-        await adb_instance.click_button(Button.exit_lobby)
-        logger.info("Queue exited due to timeout.")
-        return
-    await adb_instance.click_button(Button.accept)
-    await asyncio.sleep(2)
-
-    logger.debug("Queue accepted")
-    screenshot = await adb_instance.get_screen()
-    while screen.get_on_screen(screenshot, Image.ACCEPTED):
-        await asyncio.sleep(1)
-        screenshot = await adb_instance.get_screen()
-
-    await asyncio.sleep(3)
-
-    screenshot = await adb_instance.get_screen()
-    if screen.get_button_on_screen(screenshot, Button.accept) or screen.get_button_on_screen(screenshot, Button.play):
-        logger.debug("Queue was declined by someone else, staying in queue lock state")
-        await queue(adb_instance, config)
-
-
-_random = Random()
-
-
-async def handle_augments(screenshot: ndarray, adb_instance: ADB):
-    """
-    Checks for augments on the current screen and picks some if possible.
-
-    Args:
-        screenshot: The current screen.
-        adb_instance: The adb instance to check on.
-    """
-    is_augment_offered = screen.get_on_screen(screenshot, Image.PICK_AUGMENT)
-    if not is_augment_offered:
-        return
-
-    logger.debug("Augments offered")
-    # Roll each augment with a 50% chance
-    augment_rolls = Button.get_augment_rolls()
-    # Randomize order in which we roll
-    _random.shuffle(augment_rolls)
-    for augment in augment_rolls:
-        if bool(_random.getrandbits(1)):
-            logger.debug(f"Rolling augment offer {Button.get_augment_rolls().index(augment) + 1}")
-            await adb_instance.click_button(augment)
-        await asyncio.sleep(1)
-    await asyncio.sleep(2)
-
-    # Pick a random augment
-    augment_idx = _random.randint(0, len(Button.get_augments()) - 1)
-    augment = Button.get_augments()[augment_idx]
-    logger.debug(f"Selecting augment {augment_idx + 1}")
-    await adb_instance.click_button(augment)
-    await asyncio.sleep(1)
-
-
-async def check_surrender_state(adb_instance: ADB, screenshot: ndarray, config: AluneConfig) -> bool:
-    """
-    Check if we're able to surrender from the current game state.
-
-    Args:
-        adb_instance: The adb instance to process the surrender phase.
-        screenshot: The current screen.
-        config: An instance of the alune config to use.
-
-    Returns:
-        Whether we're able to surrender.
-    """
-    if not config.should_surrender():
-        return False
-
-    logger.debug("Checking whether we can surrender")
-    if not screen.get_on_screen(screenshot, Image.COLLAPSE_TOP_BAR):
-        await adb_instance.click_button(Button.expand_top_bar)
-        await asyncio.sleep(1)
-        screenshot = await adb_instance.get_screen()
-
-    is_phase_3_2 = screen.get_on_screen(screenshot, Image.PHASE_3_2_FULL)
-    if not is_phase_3_2:
-        return False
-
-    surrender_delay = config.get_surrender_delay()
-    logger.info(f"Surrendering the game in {surrender_delay} seconds.")
-    await asyncio.sleep(surrender_delay)
-    return True
-
-
-async def surrender_game(adb_instance: ADB):
-    """
-    Surrenders the current game.
-
-    Args:
-        adb_instance: The adb instance to process the surrender phase.
-    """
-    await adb_instance.go_back()
-    await asyncio.sleep(2)
-    await adb_instance.click_button(Button.surrender)
-    await asyncio.sleep(2)
-    await adb_instance.click_button(Button.check_surrender)
-    await asyncio.sleep(5)
-
-
-async def buy_from_shop(adb_instance: ADB, config: AluneConfig):
-    """
-    Checks the shop for traits and purchases it if found.
-
-    Args:
-        adb_instance: The adb instance to check and buy in.
-        config: An instance of the alune config to use.
-    """
-    screenshot = await adb_instance.get_screen()
-    logger.debug("Buying from shop")
-    for trait in config.get_traits():
-        search_results = screen.get_all_on_screen(
-            image=screenshot,
-            path=trait,
-            bounding_box=BoundingBox(170, 110, 1250, 230),
-            precision=0.9,
-        )
-        if len(search_results) == 0:
-            logger.debug(f"No card in the shop has the trait {trait.name}.")
-            continue
-
-        logger.debug(f"{len(search_results)} cards in the shop have the trait {trait.name}.")
-        store_cards = Button.get_store_cards()
-        _random.shuffle(store_cards)
-        for search_result in search_results:
-            for store_card in store_cards:
-                if not store_card.click_box.is_inside(search_result.get_middle()):
-                    continue
-                logger.debug(f"Buying store card {Button.get_store_cards().index(store_card) + 1}")
-                await adb_instance.click_button(store_card)
-                break
-
-            await asyncio.sleep(_random.uniform(0.25, 0.75))
-
-
-async def take_game_decision(adb_instance: ADB, config: AluneConfig):
-    """
-    Called by the game loop to take a decision in the current game.
-
-    Args:
-        adb_instance: The adb instance to take the decision in.
-        config: An instance of the alune config to use.
-    """
-    screenshot = await adb_instance.get_screen()
-
-    is_on_other_board = screen.get_button_on_screen(screenshot, Button.return_to_board)
-    if is_on_other_board:
-        logger.debug("Is on other board, checking if we are on a carousel")
-        await adb_instance.click_button(Button.return_to_board)
-        await asyncio.sleep(1)
-
-        screenshot = await adb_instance.get_screen()
-        is_in_carousel = screen.get_on_screen(screenshot, Image.CAROUSEL)
-        if is_in_carousel:
-            logger.debug("Is on carousel, clicking a random point within bounds")
-            await adb_instance.click_button(Button.return_to_board)
-            await asyncio.sleep(2)
-            # Move to a random point in the carousel area
-            await adb_instance.click_bounding_box(BoundingBox(420, 180, 825, 425))
-        return
-
-    await handle_augments(screenshot, adb_instance)
-
-    is_choose_one_hidden = screen.get_button_on_screen(screenshot, Button.choose_one_hidden, precision=0.9)
-    if is_choose_one_hidden:
-        logger.debug("Choose one is hidden, clicking it to show offers")
-        await adb_instance.click_button(Button.choose_one_hidden)
-        await asyncio.sleep(2)
-        screenshot = await adb_instance.get_screen()
-
-    is_choose_one_active = screen.get_button_on_screen(screenshot, Button.choose_one, precision=0.9)
-    if is_choose_one_active:
-        logger.debug("Choosing from an item or a choice offer")
-        await adb_instance.click_button(Button.choose_one)
-        await asyncio.sleep(1)
-        return
-
-    can_buy_xp = screen.get_button_on_screen(screenshot, Button.buy_xp)
-    if can_buy_xp and _random.randint(1, 100) <= config.get_chance_to_buy_xp():
-        logger.debug("Buying XP")
-        await adb_instance.click_button(Button.buy_xp)
-        await asyncio.sleep(1)
-
-    await buy_from_shop(adb_instance, config)
-
-    if await check_surrender_state(adb_instance, screenshot, config):
-        await surrender_game(adb_instance)
+from alune.tft.app import loop, setup_hotkeys
 
 
 async def loop_disconnect_wrapper(adb_instance: ADB, config: AluneConfig):
@@ -297,148 +42,6 @@ async def loop_disconnect_wrapper(adb_instance: ADB, config: AluneConfig):
             raise_and_exit("Could not reconnect. Please check your emulator for any errors. Exiting.")
         logger.info("Reconnected to device, continuing main loop.")
         await loop_disconnect_wrapper(adb_instance, config)
-
-
-async def take_app_decision(game_state_image_result: GameStateImageResult, adb_instance: ADB, config: AluneConfig):
-    """
-    Match the game state and take a decision based on it.
-
-    Args:
-         game_state_image_result: The result of the game state image search.
-         adb_instance: The ADB instance to take action in.
-         config: An instance of the alune config.
-    """
-    match game_state_image_result.game_state:
-        case GameState.LOADING:
-            logger.info("App state is loading...")
-            # TODO Check if the log-in prompt is on screen
-            await asyncio.sleep(10)
-        case GameState.MAIN_MENU:
-            logger.info("App state is main menu, clicking 'Play'.")
-            await adb_instance.click_button(Button.play)
-        case GameState.CHOICE_CONFIRM:
-            logger.info("App state is choice confirm, accepting the choice.")
-            await adb_instance.click_button(Button.check_choice)
-        case GameState.CHOOSE_MODE:
-            logger.info(f"App state is choose mode, selecting {config.get_game_mode()}.")
-            await adb_instance.click_image(game_state_image_result.image_result)
-        case GameState.QUEUE_MISSED:
-            logger.info("App state is queue missed, clicking it.")
-            await adb_instance.click_button(Button.check)
-        case GameState.LOBBY:
-            logger.info("App state is in lobby, locking bot into queue logic.")
-            await adb_instance.click_button(Button.play)
-            await queue(adb_instance, config)
-            logger.info("Queue lock released, likely loading into game now.")
-        case GameState.IN_GAME:
-            logger.info("App state is in game, looping decision making and waiting for the exit button.")
-            screenshot = await adb_instance.get_screen()
-            search_result = screen.get_button_on_screen(screenshot, Button.exit_now)
-            while not search_result:
-                if PAUSE_LOGIC:
-                    await asyncio.sleep(5)
-                    continue
-
-                if not await adb_instance.is_tft_active():
-                    logger.info("TFT got closed, starting it again.")
-                    await adb_instance.start_tft_app()
-                    await asyncio.sleep(10)
-                    break
-
-                await take_game_decision(adb_instance, config)
-                await asyncio.sleep(5)
-                screenshot = await adb_instance.get_screen()
-
-                game_state = await get_game_state(screenshot, config)
-                if game_state and game_state.game_state in {
-                    GameState.POST_GAME,
-                    GameState.POST_GAME_DAWN_OF_HEROES,
-                }:
-                    break
-
-                search_result = screen.get_button_on_screen(screenshot, Button.exit_now)
-            await adb_instance.click_button(Button.exit_now)
-            await asyncio.sleep(10)
-        case GameState.POST_GAME_DAWN_OF_HEROES:
-            logger.info("App state is after a game for dawn of heroes, clicking 'Continue'.")
-            await adb_instance.click_button(Button.dawn_of_heroes_continue)
-        case GameState.POST_GAME:
-            logger.info("App state is post game, clicking 'Play again'.")
-            await adb_instance.click_button(Button.play)
-
-
-async def loop(adb_instance: ADB, config: AluneConfig):
-    """
-    The main app loop logic.
-
-    Args:
-        adb_instance: An instance of the ADB connection to click in.
-        config: An instance of the alune config to use.
-    """
-    while True:
-        await delay_next_game()
-
-        if PAUSE_LOGIC:
-            await asyncio.sleep(5)
-            continue
-
-        if not await adb_instance.is_tft_active():
-            logger.info("TFT was not in the foreground, setting it as active.")
-            await adb_instance.start_tft_app()
-            await asyncio.sleep(5)
-
-        screenshot = await adb_instance.get_screen()
-        game_state_image_result = await get_game_state(screenshot, config)
-
-        if not game_state_image_result:
-            await asyncio.sleep(2)
-            continue
-
-        await take_app_decision(game_state_image_result, adb_instance, config)
-
-        await asyncio.sleep(2)
-
-
-# pylint: disable-next=too-many-return-statements
-async def get_game_state(screenshot: ndarray, config: AluneConfig) -> GameStateImageResult | None:
-    """
-    Get the current app/game state based off a screenshot.
-
-    Args:
-        screenshot: A screenshot that was taken by :class:`alune.adb.ADB`
-        config: An instance of the alune config to use.
-    """
-    if screen.get_button_on_screen(screenshot, Button.check_choice):
-        return GameStateImageResult(GameState.CHOICE_CONFIRM)
-
-    if screen.get_on_screen(screenshot, Image.RITO_LOGO):
-        return GameStateImageResult(GameState.LOADING)
-
-    if screen.get_on_screen(screenshot, Button.play.image_path) and not screen.get_on_screen(screenshot, Image.BACK):
-        return GameStateImageResult(GameState.MAIN_MENU)
-
-    game_mode_image = Image.DAWN_OF_HEROES if config.get_game_mode() == "dawn of heroes" else Image.NORMAL_GAME
-    if image_result := screen.get_on_screen(screenshot, game_mode_image):
-        return GameStateImageResult(game_state=GameState.CHOOSE_MODE, image_result=image_result)
-
-    if screen.get_button_on_screen(screenshot, Button.check):
-        return GameStateImageResult(GameState.QUEUE_MISSED)
-
-    if screen.get_on_screen(screenshot, Image.CLOSE_LOBBY) and screen.get_button_on_screen(screenshot, Button.play):
-        return GameStateImageResult(GameState.LOBBY)
-
-    if screen.get_on_screen(screenshot, Image.COMPOSITION) or screen.get_on_screen(screenshot, Image.ITEMS):
-        return GameStateImageResult(GameState.IN_GAME)
-
-    if screen.get_on_screen(screenshot, Image.BACK) and screen.get_button_on_screen(
-        screenshot, Button.dawn_of_heroes_continue
-    ):
-        return GameStateImageResult(GameState.POST_GAME_DAWN_OF_HEROES)
-
-    if screen.get_on_screen(screenshot, Image.FIRST_PLACE) and screen.get_on_screen(screenshot, Image.BACK):
-        return GameStateImageResult(GameState.POST_GAME)
-
-    return None
 
 
 async def check_phone_preconditions(adb_instance: ADB):
@@ -491,7 +94,7 @@ async def check_phone_preconditions(adb_instance: ADB):
         await adb_instance.start_tft_app()
 
 
-async def check_version():
+async def check_alune_version():
     """
     Checks the remote version against the local version and prints out a warning if remote is newer.
     """
@@ -538,7 +141,7 @@ async def main():
             level=config.get_log_level(),
         )
 
-    await check_version()
+    await check_alune_version()
 
     adb_instance = ADB(config)
 
@@ -584,57 +187,6 @@ async def main():
         )
         adb_instance.mark_screen_record_for_close()
         await asyncio.sleep(10)
-
-
-async def delay_next_game():
-    """
-    Checks whether to delay the next game based on the PLAY_NEXT_GAME variable
-    """
-    wait_counter = 0
-    while not PLAY_NEXT_GAME:
-        sleep_time = 15
-        # Don't print it every iteration
-        if wait_counter > 0 and (sleep_time * wait_counter) % 30 == 0:
-            logger.debug(f"Play next game still disabled after {sleep_time * wait_counter} seconds")
-        await asyncio.sleep(sleep_time)
-        wait_counter = wait_counter + 1
-
-
-def toggle_pause() -> None:
-    """
-    Toggles whether the bots logic evaluation should pause.
-    *Note:* This does not entirely stop the bot, but does stop various state changes that can be annoying if you're
-    trying to manually interact with it.
-    """
-    global PAUSE_LOGIC  # pylint: disable=global-statement
-    logger.debug(f"alt+p pressed, toggling pause from {PAUSE_LOGIC} to {not PAUSE_LOGIC}")
-    PAUSE_LOGIC = not PAUSE_LOGIC
-    if PAUSE_LOGIC:
-        logger.warning("Bot now paused, remember to unpause to continue botting!")
-    else:
-        logger.warning("Bot playing again!")
-
-
-def toggle_play_next_game() -> None:
-    """
-    Toggles whether the bots logic evaluation should start a new game after this finishes.
-    *Note:* This does not entirely stop the bot, but will stop it from starting a new game.
-    """
-    global PLAY_NEXT_GAME  # pylint: disable=global-statement
-    logger.debug(f"alt+n pressed, toggling pause from {PLAY_NEXT_GAME} to {not PLAY_NEXT_GAME}")
-    PLAY_NEXT_GAME = not PLAY_NEXT_GAME
-    if not PLAY_NEXT_GAME:
-        logger.warning("Bot will not queue a new game when a lobby is detected!")
-    else:
-        logger.warning("Bot will queue a new game when in lobby!")
-
-
-def setup_hotkeys() -> None:
-    """
-    Setup hotkey listeners
-    """
-    keyboard.add_hotkey("alt+p", lambda: toggle_pause())  # pylint: disable=unnecessary-lambda
-    keyboard.add_hotkey("alt+n", lambda: toggle_play_next_game())  # pylint: disable=unnecessary-lambda
 
 
 if __name__ == "__main__":
